# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  project.Member:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      joinedAt:
        type: string
      projectId:
        type: integer
      role:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  project.Project:
    properties:
      activeEpochId:
        type: integer
      completedAt:
        type: string
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      ownerUserId:
        type: integer
      progressPercent:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      visibilityMode:
        type: string
    type: object
  project.createMemberRequest:
    properties:
      role:
        type: string
      userId:
        type: integer
    required:
    - userId
    type: object
  project.createProjectRequest:
    properties:
      activeEpochId:
        type: integer
      completedAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      key:
        type: string
      name:
        type: string
      ownerUserId:
        type: integer
      progressPercent:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      visibilityMode:
        type: string
    required:
    - key
    - name
    - ownerUserId
    - status
    - visibilityMode
    type: object
  project.deleteResponse:
    properties:
      deleted:
        example: true
        type: boolean
    type: object
  project.errorResponse:
    properties:
      code:
        example: bad_request
        type: string
      message:
        example: validation failed
        type: string
    type: object
  project.updateMemberRequest:
    properties:
      projectId:
        type: integer
      role:
        type: string
      userId:
        type: integer
    type: object
  project.updateProjectRequest:
    properties:
      activeEpochId:
        type: integer
      completedAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      key:
        type: string
      name:
        type: string
      ownerUserId:
        type: integer
      progressPercent:
        type: integer
      startedAt:
        type: string
      status:
        type: string
      visibilityMode:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Seamless collaboration backend.
  title: Seamless Service API
  version: 0.2.0
paths:
...
  /project-members/{projectMemberId}:
    delete:
      parameters:
      - description: Project member ID
        in: path
        name: projectMemberId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Delete project member
      tags:
      - Projects
    get:
      parameters:
      - description: Project member ID
        in: path
        name: projectMemberId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Get project member
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project member ID
        in: path
        name: projectMemberId
        required: true
        type: integer
      - description: Project member payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/project.updateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Update project member
      tags:
      - Projects
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/project.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Create project
      tags:
      - Projects
  /projects/{projectId}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Get project
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Project payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/project.updateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Update project
      tags:
      - Projects
  /projects/{projectId}/members:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.Member'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: List project members
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Project member payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/project.createMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/project.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/project.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/project.errorResponse'
      security:
      - bearerAuth: []
      summary: Create project member
      tags:
      - Projects
...
schemes:
- http
- https
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"